Improved build process on Mac OS X 10.5

+++ Source/WindowSystem/CoreGL/testWindowCoreGL.cpp
@@ -382,7 +382,7 @@ int main (int argc, char **argv)
         (CGLPixelFormatAttribute)0
     };
     CGLPixelFormatObj pixelFormatObj;
-    long numPixelFormats;
+    GLint numPixelFormats;
     CGLChoosePixelFormat(attribs, &pixelFormatObj, &numPixelFormats); 
 
     CGLContextObj contextObj;
